Module Protocol_9p_unix.Lofs9p

include Protocol_9p.Filesystem.S

type t

A filesystem. This is likely shared amongst many connections.

type connection

Server state associated with a particular connection. This should contain connection-local data such as cached authentication information and a private space of Fids

val connect : t -> Protocol_9p__Protocol_9p_info.t -> connection

Called after making a connection to initialise the per-connection state

val attach : connection -> cancel:unit Lwt.t -> Protocol_9p__Protocol_9p_request.Attach.t -> Protocol_9p__Protocol_9p_response.Attach.t Protocol_9p.Filesystem.or_error Lwt.t

val walk : connection -> cancel:unit Lwt.t -> Protocol_9p__Protocol_9p_request.Walk.t -> Protocol_9p__Protocol_9p_response.Walk.t Protocol_9p.Filesystem.or_error Lwt.t

val clunk : connection -> cancel:unit Lwt.t -> Protocol_9p__Protocol_9p_request.Clunk.t -> unit Protocol_9p.Filesystem.or_error Lwt.t

val open_ : connection -> cancel:unit Lwt.t -> Protocol_9p__Protocol_9p_request.Open.t -> Protocol_9p__Protocol_9p_response.Open.t Protocol_9p.Filesystem.or_error Lwt.t

val read : connection -> cancel:unit Lwt.t -> Protocol_9p__Protocol_9p_request.Read.t -> Protocol_9p__Protocol_9p_response.Read.t Protocol_9p.Filesystem.or_error Lwt.t

val stat : connection -> cancel:unit Lwt.t -> Protocol_9p__Protocol_9p_request.Stat.t -> Protocol_9p__Protocol_9p_response.Stat.t Protocol_9p.Filesystem.or_error Lwt.t

val create : connection -> cancel:unit Lwt.t -> Protocol_9p__Protocol_9p_request.Create.t -> Protocol_9p__Protocol_9p_response.Create.t Protocol_9p.Filesystem.or_error Lwt.t

val write : connection -> cancel:unit Lwt.t -> Protocol_9p__Protocol_9p_request.Write.t -> Protocol_9p__Protocol_9p_response.Write.t Protocol_9p.Filesystem.or_error Lwt.t

val remove : connection -> cancel:unit Lwt.t -> Protocol_9p__Protocol_9p_request.Remove.t -> unit Protocol_9p.Filesystem.or_error Lwt.t

val wstat : connection -> cancel:unit Lwt.t -> Protocol_9p__Protocol_9p_request.Wstat.t -> unit Protocol_9p.Filesystem.or_error Lwt.t

val make : string list -> t

make root construct a filesystem which corresponds to the local Unix filesystem rooted at root