Module Server9p_unix.Make

Parameters

module Log : Protocol_9p.S.LOG

module Filesystem : Protocol_9p.Filesystem.S

Signature

type t

val of_fd : Filesystem.t -> Lwt_unix.file_descr -> t

of_fd fs fd returns a server configured to accept connections on listening socket fd.

val listen : Filesystem.t -> string -> string -> t Protocol_9p.Error.t Lwt.t

listen fs proto address callback listens on the address address and prepares to serve 9P via the callback. The proto and address can be either:

unix /path/to/socket

tcp ip:port

val shutdown : t -> unit Lwt.t

shutdown t requests that the running server t is shutdown, and blocks until the shutdown is complete and resources are freed.

val serve_forever : t -> unit Protocol_9p.Error.t Lwt.t

serve_forever t starts serving 9P requests via t. This thread only returns when the server has been shutdown.